+++ CHANGELOG.md
@@ -6,14 +6,25 @@ All notable changes to this project will be documented in this file.
 
 ### Feature
 
-- add velocity unit Angstrom per seconds
-- add force unit kcal per mole Angstrom
-- add possibility to multiply and divide more than two unit with one call (see unit_mul, unit_div)
-- add static function to create a quantity from a baseValue
+#### Units
+
+- add velocity Unit Angstrom per seconds
+- add force Unit kcal per mole Angstrom
+
+#### Quantities
+
+- add possibility to multiply and divide more than two Units with one call (see unit_mul, unit_div)
+
+#### Type Traits
+
+- introducing new type_traits library
 
 ### Breaking Changes
 
 - all pre-defined units with a division like `mps` are not used like `m_per_s`
+- remove all `sub-namespaces` besides `mstd` and `details`  
+- refactoring of how new units are build (cleaning up internal helper traits)
+- rename unit to Unit and remove real_unit
 
 <!-- insertion marker -->
 ## [0.0.1](https://github.com/97gamjak/mstd/releases/tag/0.0.1) - 2025-10-04